Move peer connection logic to ethereum protocol
